Move info event classes to Mapsui namespace
