+++ Mapsui/Map.cs
@@ -14,7 +14,6 @@ using Mapsui.Fetcher;
 using Mapsui.Layers;
 using Mapsui.Limiting;
 using Mapsui.Styles;
-using Mapsui.UI;
 using Mapsui.Widgets;
 
 namespace Mapsui;
