+++ Samples/Mapsui.Samples.Eto/MainForm.cs
@@ -7,7 +7,6 @@ namespace Mapsui.Samples.Eto;
 using System;
 using System.Text;
 using System.Linq;
-using Mapsui.UI;
 using Mapsui.UI.Eto;
 using Mapsui.Logging;
 using Mapsui.Extensions;
@@ -15,6 +14,7 @@ using Mapsui.Samples.Common;
 
 using global::Eto.Forms;
 using global::Eto.Drawing;
+using Mapsui;
 
 public class MainForm : Form
 {
