+++ Samples/Mapsui.Samples.Uno.WinUI/Mapsui.Samples.Uno.WinUI.Shared/MainPage.xaml.cs
@@ -11,7 +11,6 @@ using Mapsui.Samples.Common;
 using Mapsui.Samples.Common.Extensions;
 using Mapsui.Samples.Common.Maps;
 using Mapsui.Samples.CustomWidget;
-using Mapsui.UI;
 using Mapsui.Tiling;
 using Mapsui.UI.WinUI;
 using Mapsui.Samples.Common.Utilities;
