+++ Samples/Mapsui.Samples.Uno/Mapsui.Samples.Uno.Shared/MainPage.xaml.cs
@@ -7,7 +7,6 @@ using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
 using Mapsui.Extensions;
 using Mapsui.Providers.Wms;
-using Mapsui.UI;
 using Mapsui.Samples.Common;
 using Mapsui.Samples.Common.Extensions;
 using Mapsui.Samples.Common.Maps;
